2024, Day 9 (AI)
Part 1: ~5 shot as the initial solution was extremely slow iterating the
whole disk for each operation. After fixing that there was still a bug
not fulfilling the requirements of moving items from the end of the
disk. In the end, it was still a ~2min exection
Part 2: 1 shot but very slow (~4 min)