2024, Day 12 (AI)
Part 1: 5-6 shot. Implemented flood fill immediatelly but needed
clarifications on the perimeter.
Part 2: didn't solve even with various hinting from different angles.
I had to come up with a solution and ask to hack it leading to
this mess of a code. But in the end it worked. Also it produced
a bunch of coordinate issues (x & y are still mixed up) as well
as wrong directional vectors.